Reorganized Api felated functions into new ApiLogic.kt file
