+++ app/src/main/java/com/example/sampleappcompose/ApiLogic.kt
@@ -1,0 +1,40 @@
+package com.example.sampleappcompose
+
+import androidx.compose.runtime.MutableState
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+data class ApiResponse(val Title: String, val Year: String, val imdbRating: String)
+
+interface ApiService {
+    @GET("/")
+    suspend fun makeApiCall(@Query("apikey") apikey: String, @Query("t") title: String): ApiResponse
+}
+
+suspend fun makeApiCall(movieTitle: String): ApiResponse {
+    val retrofit = Retrofit.Builder()
+        .baseUrl("http://www.omdbapi.com/") // OMDb API
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+
+    val apiService = retrofit.create(ApiService::class.java)
+
+    return withContext(Dispatchers.IO) {
+        apiService.makeApiCall("6b938676", movieTitle)
+    }
+}
+
+suspend fun launchApiCall(movieTitle: String, apiResponse: MutableState<String>, logs: MutableList<String>) {
+    try {
+        val response = makeApiCall(movieTitle)
+        apiResponse.value = "Title: ${response.Title} \n Year: ${response.Year} \n IMDb Rating: ${response.imdbRating}"
+        //apiResponse.value = "${makeApiCall(movieTitle)}"
+        logs.add("API call made successfully.")
+    } catch (e: Exception) {
+        logs.add("API call failed: ${e.localizedMessage}")
+    }
+}
